move all consts to one file
