--- src/bridge-sync.js
@@ -1,1 +1,0 @@
-export const PACKAGE_INFO_ACTION = 'PACKAGE_INFO_ACTION';
