+++ src/consts.js
@@ -1,0 +1,4 @@
+export const PACKAGE_INFO_ACTION = 'PACKAGE_INFO_ACTION';
+export const PACKAGE_INFO_EVENT = 'get-package-info';
+export const FROM_CONTENT_SCRIPT = 'content-script';
+export const FROM_WEBPAGE = 'webpage';
