+++ src/content-to-webapp.js
@@ -1,22 +1,17 @@
+import { FROM_CONTENT_SCRIPT, FROM_WEBPAGE, PACKAGE_INFO_EVENT } from './consts';
 import { getPackageInfo } from './content/bridge';
-
-const eventType = 'get-package-info';
-const fromContentScript = 'content-script';
-const fromWebpage = 'webpage';
 
 export const listen = () => {
   window.addEventListener('message', (event) => {
-    if (event.data.from === fromWebpage) {
-      if (event.data.type === eventType) {
-        const { type, name } = event.data.payload;
-        getPackageInfo({ type, name }).then((info) => {
-          window.postMessage({
-            type: eventType,
-            from: fromContentScript,
-            payload: info,
-          });
+    if (event.data.from === FROM_WEBPAGE && event.data.type === PACKAGE_INFO_EVENT) {
+      const { type, name } = event.data.payload;
+      getPackageInfo({ type, name }).then((info) => {
+        window.postMessage({
+          type: PACKAGE_INFO_EVENT,
+          from: FROM_CONTENT_SCRIPT,
+          payload: info,
         });
-      }
+      });
     }
   });
 };
