+++ src/content/bridge.js
@@ -1,4 +1,4 @@
-import { PACKAGE_INFO_ACTION } from '../bridge-sync';
+import { PACKAGE_INFO_ACTION } from '../consts';
 
 export const getPackageInfo = async (packageID) => {
   return new Promise((resolve) =>
