+++ src/custom-elements/webapp-to-content.js
@@ -1,22 +1,16 @@
+import { PACKAGE_INFO_EVENT, FROM_CONTENT_SCRIPT, FROM_WEBPAGE } from '../consts.js';
 import * as store from './store.js';
 
-// TODO: move to shared file
-const eventType = 'get-package-info';
-const fromContentScript = 'content-script';
-const fromWebpage = 'webpage';
-
 window.addEventListener('message', (event) => {
-  if (event.data.from === fromContentScript) {
-    if (event.data.type === eventType) {
-      store.updatePackageInfo(event.data.payload);
-    }
+  if (event.data.from === FROM_CONTENT_SCRIPT && event.data.type === PACKAGE_INFO_EVENT) {
+    store.updatePackageInfo(event.data.payload);
   }
 });
 
 export const getPackageInfo = ({ type, name }) => {
   window.postMessage({
-    type: eventType,
-    from: fromWebpage,
+    type: PACKAGE_INFO_EVENT,
+    from: FROM_WEBPAGE,
     payload: { type, name },
   });
 };
